refactor(dockstat): Refactor layout into modular hooks and improve context initialization
The monolithic `layout.tsx` component has been refactored into a dedicated `src/layout` directory, organizing related logic into smaller, reusable custom hooks. This significantly improves code organization, readability, and maintainability by separating concerns.

Key changes include:
*   **Layout Component Decomposition**: The main `Layout` component is now located in `src/layout/components/Layout.tsx`.
*   **Modular Hooks**:
    *   `useLayout`: Consolidates state and effects for the primary layout.
    *   `useThemeManager`: Encapsulates theme-related logic, including application, adjustment, and creation.
    *   `useLogs`, `useRamUsage`, `usePinMutations`, `usePluginRoutes`, `useDeleteTheme`: Extract specific functionalities into dedicated hooks.
*   **QueryClient Context Initialization**: The `QueryClientContext` now initializes with a proper `new QueryClient()` instance, ensuring a fully functional client is always provided. The `QueryClientProvider` also now wraps its children with `QueryClientContext.Provider`.
*   **Theme Provider Simplification**:
    *   The `getAllThemes` function has been removed from `ThemeProviderData` and the `ThemeProvider` context, as theme list updates are now handled more directly via `useEdenQuery`.
    *   Toast notifications related to theme cache updates from theme creation mutations have been removed, aligning with the simplified theme list refresh.
