+++ apps/dockstat/src/contexts/theme.ts
@@ -19,7 +19,6 @@ export type ThemeProviderData<TDat = undefined, TInput = ThemeContextData> = {
   applyTheme: (themeName: string) => Promise<void>
   applyThemeById: (themeId: number) => Promise<void>
   themesList: ThemeListItem[] | null
-  getAllThemes: () => Promise<void>
   adjustCurrentTheme: (vars: ThemeContextData["vars"]) => void
   createNewThemeFromCurrent: MutationResult<TDat, TInput>
 }
@@ -32,7 +31,6 @@ export const ThemeProviderContext = createContext<ThemeProviderData<unknown, The
   applyTheme: async () => {},
   applyThemeById: async () => {},
   themesList: null,
-  getAllThemes: async () => {},
   adjustCurrentTheme: () => {},
   createNewThemeFromCurrent: {} as MutationResult<undefined, ThemeContextData>,
 })
