+++ apps/dockstat/src/hooks/useEdenQuery.ts
@@ -1,5 +1,5 @@
 import { extractEdenError } from "@dockstat/utils"
-import { useQuery, type UseQueryOptions } from "@tanstack/react-query"
+import { type UseQueryOptions, useQuery } from "@tanstack/react-query"
 
 type EdenQueryRoute = (options?: {
   fetch?: RequestInit
