+++ apps/dockstat/src/layout/hooks/useRamUsage.ts
@@ -1,0 +1,8 @@
+import { useEffect, useState } from "react"
+import { rssFeedEffect } from "@WSS"
+
+export function useRamUsage() {
+  const [ramUsage, setRamUsage] = useState<string>("Connecting...")
+  useEffect(() => rssFeedEffect(setRamUsage), [])
+  return ramUsage
+}
